+++ src/ts/export/index.ts
@@ -70,8 +70,8 @@ export const exportHTML = (vditor: IVditor) => {
     Vditor.mermaidRender(previewElement, '${vditor.options.cdn}', '${vditor.options.theme}');
     Vditor.flowchartRender(previewElement, '${vditor.options.cdn}');
     Vditor.graphvizRender(previewElement, '${vditor.options.cdn}');
-    Vditor.chartRender(previewElement, '${vditor.options.cdn}');
-    Vditor.mindmapRender(previewElement, '${vditor.options.cdn}');
+    Vditor.chartRender(previewElement, '${vditor.options.cdn}', '${vditor.options.theme}');
+    Vditor.mindmapRender(previewElement, '${vditor.options.cdn}', '${vditor.options.theme}');
     Vditor.abcRender(previewElement, '${vditor.options.cdn}');
     Vditor.mediaRender(previewElement);
 </script></body></html>`;